Hoist requirement data to the root Dashboard.vue
