+++ src/components/Requirements.vue
@@ -118,7 +118,6 @@ export type ShowAllCourses = {
 };
 
 type Data = {
-  reqs: readonly SingleMenuRequirement[];
   displayedMajorIndex: number;
   displayedMinorIndex: number;
   numOfColleges: number;
@@ -145,15 +144,12 @@ export default Vue.extend({
     user: Object as PropType<AppUser>,
     compact: Boolean,
     startTour: Boolean,
-  },
-  mounted() {
-    this.recomputeRequirements();
+    reqs: Array as PropType<readonly SingleMenuRequirement[]>,
   },
   data(): Data {
     return {
       displayedMajorIndex: 0,
       displayedMinorIndex: 0,
-      reqs: [],
       numOfColleges: 1,
       showAllCourses: { name: '', courses: [] },
       shouldShowAllCourses: false,
@@ -167,11 +163,6 @@ export default Vue.extend({
       tour.oncomplete(() => {
         this.$emit('showTourEndWindow');
       });
-    },
-    toggleableRequirementChoices: {
-      handler() {
-        this.recomputeRequirements();
-      },
     },
   },
   computed: {
@@ -198,58 +189,6 @@ export default Vue.extend({
     },
   },
   methods: {
-    recomputeRequirements(): void {
-      const groups = computeRequirements(
-        this.getCourseCodesArray(),
-        this.toggleableRequirementChoices,
-        this.user.college,
-        this.user.major,
-        this.user.minor
-      );
-      // Turn result into data readable by requirements menu
-      const singleMenuRequirements = groups.map(group => {
-        const singleMenuRequirement: SingleMenuRequirement = {
-          ongoing: [],
-          completed: [],
-          name: `${
-            group.groupName.charAt(0) + group.groupName.substring(1).toLowerCase()
-          } Requirements`,
-          group: group.groupName.toUpperCase() as 'COLLEGE' | 'MAJOR' | 'MINOR',
-          specific: group.specific,
-        };
-        group.reqs.forEach(req => {
-          // Create progress bar with requirement with progressBar = true
-          if (req.requirement.progressBar) {
-            singleMenuRequirement.type = this.getRequirementTypeDisplayName(
-              req.requirement.fulfilledBy
-            );
-            singleMenuRequirement.fulfilled = req.totalCountFulfilled || req.minCountFulfilled;
-            singleMenuRequirement.required =
-              (req.requirement.fulfilledBy !== 'self-check' && req.totalCountRequired) ||
-              req.minCountRequired;
-          }
-          // Default display value of false for all requirement lists
-          const displayableRequirementFulfillment = { ...req, displayDescription: false };
-          if (!req.minCountFulfilled || req.minCountFulfilled < req.minCountRequired) {
-            singleMenuRequirement.ongoing.push(displayableRequirementFulfillment);
-          } else {
-            singleMenuRequirement.completed.push(displayableRequirementFulfillment);
-          }
-        });
-        // Make number of requirements items progress bar in absense of identified progress metric
-        if (!singleMenuRequirement.type) {
-          singleMenuRequirement.type = 'Requirements';
-          singleMenuRequirement.fulfilled = singleMenuRequirement.completed.length;
-          singleMenuRequirement.required =
-            singleMenuRequirement.ongoing.length + singleMenuRequirement.completed.length;
-        }
-        return singleMenuRequirement;
-      });
-      this.reqs = singleMenuRequirements;
-    },
-    getRequirementTypeDisplayName(type: string): string {
-      return type.charAt(0).toUpperCase() + type.substring(1);
-    },
     showMajorOrMinorRequirements(id: number, group: string) {
       if (group === 'MAJOR') {
         return id === this.displayedMajorIndex + this.numOfColleges;
@@ -266,23 +205,6 @@ export default Vue.extend({
         [requirementID]: option,
       };
       this.$emit('on-toggleable-requirement-choices-change', newToggleableRequirementChoices);
-    },
-    getCourseCodesArray(): readonly CourseTaken[] {
-      const courses: CourseTaken[] = [];
-      this.semesters.forEach(semester => {
-        semester.courses.forEach(course => {
-          courses.push({
-            code: `${course.lastRoster}: ${course.subject} ${course.number}`,
-            subject: course.subject,
-            courseId: course.crseId,
-            number: course.number,
-            credits: course.credits,
-            roster: course.lastRoster,
-          });
-        });
-      });
-      courses.push(...getCourseEquivalentsFromUserExams(this.user));
-      return courses;
     },
     activateMajor(id: number) {
       this.displayedMajorIndex = id;
